refactor(app:cns-website): Tweak landing page
