+++ apps/cns-website/src/app/app.routes.ts
@@ -3,7 +3,7 @@ import { ContentPageDataSchema } from '@hra-ui/design-system/content-templates/c
 import { createYamlSpecResolver } from '@hra-ui/design-system/content-templates/resolvers';
 import { NotFoundPageComponent } from '@hra-ui/design-system/error-pages/not-found-page';
 import { ContentPageComponent } from './components/content-page/content-page.component';
-import { LandingPageComponent } from './pages/landing-page.component';
+import { LandingPageComponent } from './pages/landing-page/landing-page.component';
 import { PeopleProfileComponent } from './pages/people-profile/people-profile.component';
 
 /** Application routes */
