+++ apps/cns-website/src/app/pages/landing-page/landing-page.component.scss
@@ -13,7 +13,7 @@
     align-items: center;
     flex-grow: 1;
     padding: 6.25rem 2.5rem;
-    background-image: url(../../../public/assets/bg-homepage.png);
+    background-image: url(../../../../public/assets/bg-homepage.png);
     background-size: cover;
     background-repeat: no-repeat;
     background-position: center;
