+++ libs/design-system/cards/content-button/src/index.ts
@@ -1,2 +1,2 @@
 export { ContentButtonComponent } from './lib/content-button.component';
-export { ContentButtonSchema } from './lib/content-button.schema';
+export { ContentButton, ContentButtonSchema } from './lib/content-button.schema';
