+++ libs/design-system/cards/content-button/src/lib/content-button.schema.ts
@@ -7,7 +7,7 @@ export type ContentButton = z.infer<typeof ContentButtonSchema>;
 /** Schema for content button action card data */
 export const ContentButtonSchema = ContentTemplateSchema.extend({
   component: z.literal('ContentButton'),
-  image: z.string(),
+  imageSrc: z.string(),
   date: z.string(),
   tagline: z.string(),
   tags: z.array(z.string()).optional(),
